fix module organization and offer module to offer just one to homework

+++ src/offer/offer.service.ts
@@ -17,11 +17,9 @@ export class OfferService {
     offerDto: OfferDto,
     user: User,
   ): Promise<Offer> {
-    console.log(idHomework);
     const getHomeWork = await this.HomeworkRepository.findOne(idHomework);
-    console.log(getHomeWork);
     if (!getHomeWork) {
-      throw new Error('Homework not found');
+      throw new InternalServerErrorException('Homework not found');
     }
     return this.offerRepository.makeOffer(getHomeWork, user, offerDto);
   }
@@ -29,16 +27,40 @@ export class OfferService {
   async getOffersByHomeworks(idHomework: string): Promise<Offer[]> {
     const getHomeWork = await this.HomeworkRepository.findOne(idHomework);
     if (!getHomeWork) {
-      throw new Error('Homework not found');
-      /* return InternalServerErrorException('Homework not found'); */
+      throw new InternalServerErrorException('Homework not found');
     }
     return this.offerRepository.getOffersByHomeworks(getHomeWork);
   }
   async deleteOffer(user: User, idOffer: string): Promise<Offer> {
     const getOffer = await this.offerRepository.findOne(idOffer);
     if (!getOffer) {
-      throw new Error('Homework not found');
+      throw new InternalServerErrorException('Offer not found');
+    }
+    if (getOffer.user.id !== user.id) {
+      throw new InternalServerErrorException(
+        'You are not the owner of this offer',
+      );
     }
     return this.offerRepository.deleteOffer(user, idOffer);
+  }
+  async editOffer(
+    user: User,
+    idOffer: string,
+    offerDto: OfferDto,
+  ): Promise<Offer> {
+    const getOffer = await this.offerRepository.findOne(idOffer);
+    if (!getOffer) {
+      throw new InternalServerErrorException('Homework not found');
+    }
+    if (getOffer.user.id !== user.id) {
+      throw new InternalServerErrorException(
+        'You are not the owner of this offer',
+      );
+    }
+    return await this.offerRepository.save({
+      ...getOffer,
+      ...offerDto,
+      edited: true,
+    });
   }
 }
